[DF-2617] Improved performance of events in scene.

+++ Sources/Internal/Base/TemplateHelpers.h
@@ -175,7 +175,7 @@ public:
         if (!c)
         {//assert emulation )
             int i = 0;
-            *((int*)i) = 0;
+            *((int*)i) = (int)0;
         }
         return c;
 #else
